Add a bench test with positions stolen from stockfish

+++ Charlie3/MoveInfo.cs
@@ -8,10 +8,10 @@ namespace Charlie3
         public IEnumerable<Move> Moves { get; }
         public int Evaluation { get; }
         public bool IsMate { get; }
-        public uint Time { get; }
+        public long Time { get; }
         public ulong Nodes { get; }
 
-        public MoveInfo(int depth, IEnumerable<Move> moves, int evaluation, bool isMate, uint time, ulong nodes) =>
+        public MoveInfo(int depth, IEnumerable<Move> moves, int evaluation, bool isMate, long time, ulong nodes) =>
             (Depth, Moves, Evaluation, IsMate, Time, Nodes) = (depth, moves, evaluation, isMate, time, nodes);
     }
 }
